From Mathias Froehlich, "I have today built svn trunk on HP-UX and Solaris8. I had to do two small changes to src/osgWidget/Input.cpp and src/osgAnimation/Animation.cpp mainly because of name lookup problems. "

+++ src/osgAnimation/Animation.cpp
@@ -112,7 +112,7 @@ bool Animation::update (float time)
         if (!_duration)
             t = _startTime;
         else if (t > _duration)
-            t = fmodf(t, _duration);
+            t = fmod(t, (float)_duration);
         //      std::cout << "t " << t << " duration " << _duration << std::endl;
         break;
     case PPONG: 
@@ -121,7 +121,7 @@ bool Animation::update (float time)
         else 
         {
             int tt = (int) (t / _duration);
-            t = fmodf(t, _duration);
+            t = fmod(t, (float)_duration);
             if (tt%2)
                 t = _duration - t;
         }
